x/evm: split keeper.go
